Wasgo-BE/requirements_waste_management.txt
# Waste Management Specific Dependencies
# These packages are required for the WasteBin and WasteProvider apps

# Core GIS and Geospatial Libraries
djangorestframework-gis>=1.2.0
GDAL>=3.0.0
geopy>=2.3.0
shapely>=1.8.0

# Additional Utilities
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Real-time Communication (for IoT sensors)
channels>=4.0.0
channels-redis>=4.0.0

# Task Queue (for background processing)
celery>=5.2.0
redis>=4.0.0

# Data Visualization (for analytics)
matplotlib>=3.5.0
seaborn>=0.11.0

# API Documentation
drf-spectacular>=0.22.0

# Testing
pytest>=6.2.0
pytest-django>=4.5.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
